backend/requirements.txt
requests
pydantic
python-dotenv
fastapi
uvicorn
Markdown
markdownify
google-generativeai
